refactor: 타입과 초기값 상수 분리
