+++ hooks/src/lib/constants/cardNumber.ts
@@ -1,0 +1,6 @@
+export const initialCardNumberState = {
+  first: '',
+  second: '',
+  third: '',
+  fourth: ''
+};
