+++ hooks/src/lib/constants/expiration.ts
@@ -1,0 +1,4 @@
+export const initialExpirationState = {
+  year: '',
+  month: ''
+};
