+++ hooks/src/lib/types/cardNumber.ts
@@ -1,0 +1,5 @@
+import { initialCardNumberState } from '../constants/cardNumber';
+
+export type CarNumberStateType = typeof initialCardNumberState;
+
+export type CardNumberStateKey = keyof CarNumberStateType;
