+++ hooks/src/lib/types/expiration.ts
@@ -1,0 +1,4 @@
+import { initialExpirationState } from '../constants/expiration';
+
+export type ExpirationStateType = typeof initialExpirationState;
+export type ExpirationStateKey = keyof ExpirationStateType;
