+++ hooks/src/lib/validation.ts
@@ -1,39 +1,35 @@
+const MIN_MONTH = 1;
+const MAX_MONTH = 12;
+const CARD_NUMBER_LENGTH = 4;
+const CURRENT_YEAR = new Date().getFullYear() % 100;
+const REGEX_ONLY_NUMBERS = /^[0-9]+$/;
+
 const validation = {
-  isNumber: (value: string) => {
-    const regex = /^[0-9]+$/;
-    return regex.test(value);
+  isNumber: (value: string): boolean => {
+    return REGEX_ONLY_NUMBERS.test(value);
   },
-  isValidLength: (value: string, length: number) => {
+
+  isValidLength: (value: string, length: number = CARD_NUMBER_LENGTH): boolean => {
     return value.length === length;
   },
-  isValidMonth: (value: string) => {
+
+  isValidMonth: (value: string): boolean => {
     const numericMonth = Number(value);
-    if (numericMonth < 1 || numericMonth > 12) {
-      return false;
-    }
-    return true;
+    return numericMonth >= MIN_MONTH && numericMonth <= MAX_MONTH;
   },
 
-  isValidYear: (value: string) => {
+  isValidYear: (value: string): boolean => {
     const numericYear = Number(value);
-    const currentYear = new Date().getFullYear() % 100;
-
-    if (currentYear > numericYear) {
-      return false;
-    }
-
-    return true;
+    return numericYear >= CURRENT_YEAR;
   },
-  isValidateDate: (month: string, year: string) => {
+
+  isValidateDate: (month: string, year: string): boolean => {
     const numericMonth = Number(month);
     const numericYear = Number(year);
-    const currentYear = new Date().getFullYear() % 100;
 
-    if (currentYear === numericYear) {
+    if (numericYear === CURRENT_YEAR) {
       const currentMonth = new Date().getMonth() + 1;
-      if (currentMonth > numericMonth) {
-        return false;
-      }
+      return numericMonth >= currentMonth;
     }
 
     return true;
